requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1